requirements.txt
click>=8.0.0
pandas>=1.5.0
PyYAML>=6.0
rich>=12.0.0
openpyxl>=3.0.0